#5874 Merge migrations
